rename tool api
