+++ app/desktop/desktop_server.py
@@ -22,7 +22,7 @@ from app.desktop.studio_server.prompt_api import connect_prompt_api
 from app.desktop.studio_server.provider_api import connect_provider_api
 from app.desktop.studio_server.repair_api import connect_repair_api
 from app.desktop.studio_server.settings_api import connect_settings
-from app.desktop.studio_server.tool_servers_api import connect_tool_servers_api
+from app.desktop.studio_server.tool_api import connect_tool_servers_api
 from app.desktop.studio_server.webhost import connect_webhost
 
 # Loads github pages hosted JSON config.
